requirements.txt
streamlit
pytesseract
pillow
bcrypt